Add files via upload
Final Map Cluster picture using Folium for the 4 clusters under consideration for Indian Cuisines (including Himalayan Cuisines). Clearly two of the clusters are so busy and multiple in count. But two of them are minimal although it is densely populated. There is lot of scope to open the Indian Cuisine in Cluster 3 and Cluster 4.